pre_save for setting up self, post-save for calls to other models (incase they reference self)

+++ app/models/order.rb
@@ -50,6 +50,7 @@ class Order < ActiveRecord::Base
 
   before_create -> {self.status||=:cart}
   before_save :pre_save
+  after_save :post_save
 
   ## Scopes
 
@@ -143,15 +144,18 @@ class Order < ActiveRecord::Base
         self.type=((status=="cart") ? "Cart" : nil)
         #Record timestamps if order changes status
         self.send(status+"_at=", DateTime.now) if self.respond_to?(status+"_at=")
-        #Release stock if an order has been cancelled
-        if status=="cancelled"
-          line_items.all?(&:release_stock)
-          redemptions.destroy_all
-        end
         #fix a copy of the costs when the order gets placed
         fix_costs if status=="placed"
       end
       true
+    end
+
+    def post_save
+      #Release stock if an order has been cancelled
+      if status_changed? and status=="cancelled"
+        line_items.all?(&:release_stock)
+        redemptions.destroy_all
+      end
     end
 
     def stock_check
